test(browser): fold test-only Story harnesses into their browser tests
With the Storybook re-exports gone, the 24 *Story.tsx files left under
__tests__ were imported by exactly one *.browser.test.tsx each, and the
Story suffix suggested a Storybook file that no story used. Move each
harness into the test that renders it, as a local FooHarness component
(FooStoryProps -> FooHarnessProps), so every *Story.tsx that remains is a
Storybook harness owned by a *.stories.tsx.

Includes CopyPasteStory/CopyPasteFieldsStory, which never had a story but
shared the naming, and renames the inline InlineComposerStory component in
the two InlineComposerMentions tests for the same reason. Where a test
imported SanityDocument from @sanity/client and its harness from
@sanity/types, the @sanity/types import is kept. DisableFocusAndUnset's
harness had kept a stale FocusTrackingStory name from a copy-paste; it now
matches the test name. Annotations gets a second schema constant because
both of its harnesses used SCHEMA_TYPES.
