+++ packages/sanity/src/core/comments-v2/components/pte/comment-input/__tests__/InlineComposerMentions.browser.test.tsx
@@ -42,7 +42,7 @@ const MENTION_DATA: UserListWithPermissionsHookValue = {
  * field root that is a single line tall, and that root is the boundary element
  * the composer inherits.
  */
-function InlineComposerStory() {
+function InlineComposerHarness() {
   const [value, setValue] = useState<PortableTextBlock[] | null>(null)
   const [fieldElement, setFieldElement] = useState<HTMLDivElement | null>(null)
   const [referenceElement, setReferenceElement] = useState<HTMLSpanElement | null>(null)
@@ -76,7 +76,7 @@ function InlineComposerStory() {
 
 describe('Inline comment composer', () => {
   it('keeps the mentions menu sized when the field boundary is one line tall (SAPP-4093)', async () => {
-    void render(<InlineComposerStory />)
+    void render(<InlineComposerHarness />)
 
     const $editable = page.getByTestId('comment-input-editable')
     await expect.element($editable).toBeVisible()
@@ -96,7 +96,7 @@ describe('Inline comment composer', () => {
   })
 
   it('does not treat picking a mention as a click outside the composer', async () => {
-    void render(<InlineComposerStory />)
+    void render(<InlineComposerHarness />)
 
     const $editable = page.getByTestId('comment-input-editable')
     await expect.element($editable).toBeVisible()
@@ -113,7 +113,7 @@ describe('Inline comment composer', () => {
 
   it('asks to discard when clicking outside the composer with a draft value', async () => {
     const {insertPortableText} = testHelpers()
-    void render(<InlineComposerStory />)
+    void render(<InlineComposerHarness />)
 
     const $editable = page.getByTestId('comment-input-editable')
     await expect.element($editable).toBeVisible()
