+++ packages/sanity/src/core/form/inputs/PortableText/__tests__/Decorators.browser.test.tsx
@@ -1,9 +1,84 @@
+import {BulbOutlineIcon} from '@sanity/icons/BulbOutline'
+import {defineArrayMember, defineField, defineType} from '@sanity/types'
 import {describe, expect, it} from 'vitest'
 import {render} from 'vitest-browser-react'
 import {page} from 'vitest/browser'
 
+import {TestForm} from '../../../../../../test/browser/TestForm'
 import {testHelpers} from '../../../../../../test/browser/testHelpers'
-import {DecoratorsStory} from './DecoratorsStory'
+import {TestWrapper} from '../../../../../../test/browser/TestWrapper'
+import {type BlockDecoratorProps} from '../../../types/blockProps'
+
+function Highlight(props: BlockDecoratorProps) {
+  return (
+    <span
+      data-testid="custom-highlight-decorator"
+      data-title={props.title}
+      data-value={props.value}
+    >
+      {props.renderDefault(props)}
+    </span>
+  )
+}
+
+function Spoiler(props: BlockDecoratorProps) {
+  return (
+    <span data-testid="custom-spoiler-decorator" style={{color: 'red'}}>
+      {props.children}
+    </span>
+  )
+}
+
+const SCHEMA_TYPES = [
+  defineType({
+    type: 'document',
+    name: 'test',
+    title: 'Test',
+    fields: [
+      defineField({
+        type: 'array',
+        name: 'defaultDecorators',
+        of: [
+          defineArrayMember({
+            type: 'block',
+          }),
+        ],
+      }),
+      defineField({
+        type: 'array',
+        name: 'customDecorator',
+        of: [
+          defineArrayMember({
+            type: 'block',
+            marks: {
+              decorators: [
+                {
+                  title: 'Highlight',
+                  value: 'highlight',
+                  icon: BulbOutlineIcon,
+                  component: Highlight,
+                },
+                {
+                  title: 'Spoiler',
+                  value: 'spoiler',
+                  component: Spoiler,
+                },
+              ],
+            },
+          }),
+        ],
+      }),
+    ],
+  }),
+]
+
+function DecoratorsHarness() {
+  return (
+    <TestWrapper schemaTypes={SCHEMA_TYPES}>
+      <TestForm />
+    </TestWrapper>
+  )
+}
 
 const DEFAULT_DECORATORS = [
   {
@@ -49,7 +124,7 @@ describe('Portable Text Input', () => {
         insertPortableText,
         toggleHotkey,
       } = testHelpers()
-      void render(<DecoratorsStory />)
+      void render(<DecoratorsHarness />)
       const $portableTextInput = await getFocusedPortableTextInput('field-defaultDecorators')
       const $pte = await getFocusedPortableTextEditor('field-defaultDecorators')
       const modifierKey = getModifierKey()
@@ -85,7 +160,7 @@ describe('Portable Text Input', () => {
     describe('Toolbar buttons', () => {
       it('Should display all default decorator buttons', async () => {
         const {getFocusedPortableTextInput} = testHelpers()
-        void render(<DecoratorsStory />)
+        void render(<DecoratorsHarness />)
         const $portableTextInput = await getFocusedPortableTextInput('field-defaultDecorators')
 
         // Assertion: All buttons in the menu bar should be visible and have icon
@@ -101,7 +176,7 @@ describe('Portable Text Input', () => {
 
       it('Should display custom decorator button and icon', async () => {
         const {getFocusedPortableTextInput} = testHelpers()
-        void render(<DecoratorsStory />)
+        void render(<DecoratorsHarness />)
         const $portableTextInput = await getFocusedPortableTextInput('field-customDecorator')
         // Assertion: Button for highlight should exist
         const $highlightButton = $portableTextInput.getByRole('button', {name: 'Highlight'})
@@ -121,7 +196,7 @@ describe('Portable Text Input', () => {
         getFocusedPortableTextEditor,
         insertPortableText,
       } = testHelpers()
-      void render(<DecoratorsStory />)
+      void render(<DecoratorsHarness />)
       const $portableTextInput = await getFocusedPortableTextInput('field-customDecorator')
       const $pte = await getFocusedPortableTextEditor('field-customDecorator')
 
@@ -155,7 +230,7 @@ describe('Portable Text Input', () => {
         getFocusedPortableTextEditor,
         insertPortableText,
       } = testHelpers()
-      void render(<DecoratorsStory />)
+      void render(<DecoratorsHarness />)
       const $portableTextInput = await getFocusedPortableTextInput('field-customDecorator')
       const $pte = await getFocusedPortableTextEditor('field-customDecorator')
