--- packages/sanity/src/core/form/inputs/PortableText/__tests__/DragAndDropStory.tsx
@@ -1,73 +1,0 @@
-import {type SanityDocument} from '@sanity/client'
-import {defineArrayMember, defineField, defineType, type Path} from '@sanity/types'
-
-import {TestForm} from '../../../../../../test/browser/TestForm'
-import {TestWrapper} from '../../../../../../test/browser/TestWrapper'
-
-const SCHEMA_TYPES = [
-  defineType({
-    type: 'document',
-    name: 'test',
-    title: 'Test',
-    fields: [
-      defineField({
-        type: 'array',
-        name: 'body',
-        of: [
-          defineArrayMember({
-            type: 'block',
-            of: [
-              defineArrayMember({
-                type: 'object',
-                name: 'inlineObjectWithTextProperty',
-                fields: [
-                  defineField({
-                    type: 'string',
-                    name: 'text',
-                    components: {
-                      input: (inputProps) => (
-                        <div data-testid="inlineTextInputField">
-                          {inputProps.renderDefault(inputProps)}
-                        </div>
-                      ),
-                    },
-                  }),
-                ],
-              }),
-            ],
-          }),
-          defineArrayMember({
-            type: 'object',
-            name: 'testObjectBlock',
-            fields: [{type: 'string', name: 'text'}],
-            components: {
-              input: (inputProps) => (
-                <div data-testid="objectBlockInputField">
-                  {inputProps.renderDefault(inputProps)}
-                </div>
-              ),
-            },
-          }),
-        ],
-      }),
-    ],
-  }),
-]
-
-function DragAndDropStory({
-  focusPath,
-  onPathFocus,
-  document,
-}: {
-  focusPath?: Path
-  onPathFocus?: (path: Path) => void
-  document?: SanityDocument
-}) {
-  return (
-    <TestWrapper schemaTypes={SCHEMA_TYPES}>
-      <TestForm document={document} focusPath={focusPath} onPathFocus={onPathFocus} />
-    </TestWrapper>
-  )
-}
-
-export default DragAndDropStory
