--- packages/sanity/src/core/form/inputs/PortableText/__tests__/InputStory.tsx
@@ -1,59 +1,0 @@
-import {type PortableTextEditor} from '@portabletext/editor'
-import {defineArrayMember, defineField, defineType} from '@sanity/types'
-import {useMemo} from 'react'
-import {type InputProps, type PortableTextInputProps} from 'sanity'
-
-import {TestForm} from '../../../../../../test/browser/TestForm'
-import {TestWrapper} from '../../../../../../test/browser/TestWrapper'
-
-interface InputStoryProps {
-  // oxlint-disable-next-line no-deprecated -- will fix in follow up PR
-  editorRef?: React.Ref<PortableTextEditor | null>
-  ptInputProps?: Partial<PortableTextInputProps>
-}
-
-export function InputStory(props: InputStoryProps) {
-  const {editorRef, ptInputProps} = props
-
-  const schemaTypes = useMemo(
-    () => [
-      defineType({
-        type: 'document',
-        name: 'test',
-        title: 'Test',
-        fields: [
-          defineField({
-            type: 'array',
-            name: 'body',
-            of: [
-              defineArrayMember({
-                type: 'block',
-              }),
-            ],
-            components: {
-              input: (inputProps: InputProps) => {
-                const editorProps = {
-                  ...inputProps,
-                  ...ptInputProps,
-                  editorRef,
-                } as PortableTextInputProps
-                return (
-                  <div data-testid="pt-input-with-editor-ref">
-                    {inputProps.renderDefault(editorProps)}
-                  </div>
-                )
-              },
-            },
-          }),
-        ],
-      }),
-    ],
-    [ptInputProps, editorRef],
-  )
-
-  return (
-    <TestWrapper schemaTypes={schemaTypes}>
-      <TestForm />
-    </TestWrapper>
-  )
-}
