+++ packages/sanity/src/core/form/inputs/PortableText/__tests__/NestedInput.browser.test.tsx
@@ -1,9 +1,115 @@
-import {type SanityDocument} from '@sanity/types'
+import {
+  defineArrayMember,
+  defineField,
+  defineType,
+  type Path,
+  type SanityDocument,
+} from '@sanity/types'
 import {describe, expect, it} from 'vitest'
 import {page, userEvent} from 'vitest/browser'
 
+import {TestForm} from '../../../../../../test/browser/TestForm'
 import {testHelpers} from '../../../../../../test/browser/testHelpers'
-import NestedInputStory from './NestedInputStory'
+import {TestWrapper} from '../../../../../../test/browser/TestWrapper'
+
+const SCHEMA_TYPES = [
+  defineType({
+    type: 'document',
+    name: 'test',
+    title: 'Test',
+    fields: [
+      defineField({
+        type: 'array',
+        name: 'body',
+        of: [
+          defineArrayMember({
+            type: 'block',
+            of: [
+              defineArrayMember({
+                type: 'object',
+                title: 'Inline Object',
+                fields: [
+                  defineField({
+                    type: 'string',
+                    name: 'title',
+                    title: 'Title',
+                  }),
+                  defineField({
+                    type: 'array',
+                    name: 'caption',
+                    title: 'Caption',
+                    of: [{type: 'block'}],
+                  }),
+                ],
+                components: {
+                  input: (inputProps) => {
+                    return (
+                      <div data-testid="inlinePopover">{inputProps.renderDefault(inputProps)}</div>
+                    )
+                  },
+                },
+              }),
+            ],
+          }),
+          defineArrayMember({
+            type: 'object',
+            name: 'nestedObjectBlock',
+            title: 'Nested Object Block',
+            fields: [
+              defineField({
+                type: 'array',
+                name: 'siteOverrides',
+                title: 'Site overrides',
+                of: [
+                  defineArrayMember({
+                    type: 'object',
+                    name: 'siteOverride',
+                    fields: [
+                      defineField({
+                        type: 'array',
+                        name: 'content',
+                        title: 'Content',
+                        of: [
+                          {
+                            type: 'block',
+                            marks: {
+                              annotations: [
+                                {
+                                  type: 'object',
+                                  name: 'nestedAnnotation',
+                                  title: 'Nested annotation',
+                                  fields: [
+                                    defineField({
+                                      type: 'string',
+                                      name: 'value',
+                                      title: 'Annotation value',
+                                    }),
+                                  ],
+                                },
+                              ],
+                            },
+                          },
+                        ],
+                      }),
+                    ],
+                  }),
+                ],
+              }),
+            ],
+          }),
+        ],
+      }),
+    ],
+  }),
+]
+
+function NestedInputHarness({document, focusPath}: {document?: SanityDocument; focusPath?: Path}) {
+  return (
+    <TestWrapper schemaTypes={SCHEMA_TYPES}>
+      <TestForm document={document} focusPath={focusPath} />
+    </TestWrapper>
+  )
+}
 
 const {render} = await import('vitest-browser-react')
 
@@ -16,7 +122,7 @@ describe('Portable Text Input', () => {
         waitForFocusedNodeText,
         waitForSelectionOffsets,
       } = testHelpers()
-      void render(<NestedInputStory />)
+      void render(<NestedInputHarness />)
 
       const $portableTextInput = await getFocusedPortableTextInput('field-body')
 
@@ -110,7 +216,7 @@ describe('Portable Text Input', () => {
         ],
       }
       void render(
-        <NestedInputStory
+        <NestedInputHarness
           document={document}
           focusPath={[
             'body',
