--- packages/sanity/src/core/form/inputs/PortableText/__tests__/PresenceCursorsStory.tsx
@@ -1,39 +1,0 @@
-import {defineArrayMember, defineField, defineType, type SanityDocument} from '@sanity/types'
-import {type FormNodePresence} from 'sanity'
-
-import {TestForm} from '../../../../../../test/browser/TestForm'
-import {TestWrapper} from '../../../../../../test/browser/TestWrapper'
-
-const schemaTypes = [
-  defineType({
-    type: 'document',
-    name: 'test',
-    title: 'Test',
-    fields: [
-      defineField({
-        type: 'array',
-        name: 'body',
-        of: [
-          defineArrayMember({
-            type: 'block',
-          }),
-        ],
-      }),
-    ],
-  }),
-]
-
-interface PresenceCursorsStoryProps {
-  presence: FormNodePresence[]
-  document: SanityDocument
-}
-
-export function PresenceCursorsStory(props: PresenceCursorsStoryProps) {
-  const {document, presence} = props
-
-  return (
-    <TestWrapper schemaTypes={schemaTypes}>
-      <TestForm document={document} presence={presence} />
-    </TestWrapper>
-  )
-}
