--- packages/sanity/src/core/form/inputs/PortableText/__tests__/StylesStory.tsx
@@ -1,40 +1,0 @@
-import {defineArrayMember, defineField, defineType, type SanityDocument} from '@sanity/types'
-
-import {TestForm} from '../../../../../../test/browser/TestForm'
-import {TestWrapper} from '../../../../../../test/browser/TestWrapper'
-
-const SCHEMA_TYPES = [
-  defineType({
-    type: 'document',
-    name: 'test',
-    title: 'Test',
-    fields: [
-      defineField({
-        type: 'array',
-        name: 'defaultStyles',
-        of: [
-          defineArrayMember({
-            type: 'block',
-          }),
-        ],
-      }),
-      defineField({
-        type: 'array',
-        name: 'oneStyle',
-        of: [
-          defineArrayMember({
-            type: 'block',
-            styles: [{title: 'Normal', value: 'normal'}],
-          }),
-        ],
-      }),
-    ],
-  }),
-]
-export function StylesStory(props: {document?: SanityDocument}) {
-  return (
-    <TestWrapper schemaTypes={SCHEMA_TYPES}>
-      <TestForm document={props.document} />
-    </TestWrapper>
-  )
-}
