+++ packages/sanity/src/core/form/inputs/PortableText/__tests__/TableValidationDepth.browser.test.tsx
@@ -1,8 +1,105 @@
-import {type SanityDocument} from '@sanity/types'
+import {
+  defineArrayMember,
+  defineField,
+  defineType,
+  type PortableTextBlock,
+  type SanityDocument,
+} from '@sanity/types'
 import {describe, expect, it} from 'vitest'
 
+import {TestForm} from '../../../../../../test/browser/TestForm'
 import {testHelpers} from '../../../../../../test/browser/testHelpers'
-import {TableValidationDepthStory} from './TableValidationDepthStory'
+import {TestWrapper} from '../../../../../../test/browser/TestWrapper'
+
+// The same custom rule guards root blocks and cell blocks, so validation
+// marker rendering can be compared across depths.
+const noBadWords = (block: PortableTextBlock | undefined) => {
+  const hasBad =
+    Array.isArray(block?.children) &&
+    block.children.some((child) => typeof child.text === 'string' && child.text.includes('bad'))
+  return hasBad ? 'No bad words' : true
+}
+
+const SCHEMA_TYPES = [
+  defineType({
+    type: 'document',
+    name: 'test',
+    title: 'Test',
+    fields: [
+      defineField({
+        type: 'array',
+        name: 'body',
+        of: [
+          defineArrayMember({
+            type: 'block',
+            validation: (Rule) => Rule.custom(noBadWords),
+          }),
+          defineArrayMember({
+            type: 'object',
+            name: 'table',
+            fields: [
+              defineField({type: 'number', name: 'headerRows'}),
+              defineField({
+                type: 'array',
+                name: 'rows',
+                of: [
+                  defineArrayMember({
+                    type: 'object',
+                    name: 'row',
+                    fields: [
+                      defineField({
+                        type: 'array',
+                        name: 'cells',
+                        of: [
+                          defineArrayMember({
+                            type: 'object',
+                            name: 'cell',
+                            fields: [
+                              defineField({
+                                type: 'array',
+                                name: 'value',
+                                of: [
+                                  defineArrayMember({
+                                    type: 'block',
+                                    validation: (Rule) => Rule.custom(noBadWords),
+                                  }),
+                                ],
+                              }),
+                            ],
+                          }),
+                        ],
+                      }),
+                    ],
+                  }),
+                ],
+              }),
+            ],
+          }),
+        ],
+        components: {
+          portableText: {
+            plugins: (props) =>
+              props.renderDefault({
+                ...props,
+                plugins: {
+                  ...props.plugins,
+                  table: {enabled: true},
+                },
+              }),
+          },
+        },
+      }),
+    ],
+  }),
+]
+
+function TableValidationDepthHarness(props: {document?: SanityDocument}) {
+  return (
+    <TestWrapper schemaTypes={SCHEMA_TYPES}>
+      <TestForm document={props.document} />
+    </TestWrapper>
+  )
+}
 
 const {render} = await import('vitest-browser-react')
 
@@ -45,7 +142,7 @@ describe('Portable Text Input - validation markers at depth', () => {
   it('renders the error marker on failing blocks at root and inside table cells alike', async () => {
     const {getFocusedPortableTextEditor} = testHelpers()
 
-    void render(<TableValidationDepthStory document={document} />)
+    void render(<TableValidationDepthHarness document={document} />)
 
     const $pte = await getFocusedPortableTextEditor('field-body')
     await expect.element($pte).toHaveTextContent('bad cell text')
