+++ packages/sanity/src/core/form/inputs/PortableText/__tests__/copyPaste/CopyPaste.browser.test.tsx
@@ -1,10 +1,17 @@
-import {type Path, type SanityDocument} from '@sanity/types'
+import {
+  defineArrayMember,
+  defineField,
+  defineType,
+  type Path,
+  type SanityDocument,
+} from '@sanity/types'
 import {beforeEach, describe, expect, it} from 'vitest'
 import {render} from 'vitest-browser-react'
 import {page, server} from 'vitest/browser'
 
+import {TestForm} from '../../../../../../../test/browser/TestForm'
 import {testHelpers} from '../../../../../../../test/browser/testHelpers'
-import CopyPasteStory from './CopyPasteStory'
+import {TestWrapper} from '../../../../../../../test/browser/TestWrapper'
 import {
   CLEANED_UNICODE_INPUT_SNAPSHOT,
   GDOCS_INPUT,
@@ -12,6 +19,78 @@ import {
   REMOVED_INPUT_SNAPSHOT,
   UNICODE_TEXT,
 } from './input'
+
+const SCHEMA_TYPES = [
+  defineType({
+    type: 'document',
+    name: 'test',
+    title: 'Test',
+    fields: [
+      defineField({
+        type: 'array',
+        name: 'body',
+        of: [
+          defineArrayMember({
+            type: 'block',
+            options: {
+              unstable_whitespaceOnPasteMode: 'remove',
+            },
+          }),
+          defineArrayMember({
+            type: 'image',
+            name: 'image',
+            title: 'Image block',
+            preview: {
+              select: {
+                fileName: 'asset.originalFilename',
+                image: 'asset',
+              },
+              prepare({fileName, image}) {
+                return {
+                  media: image,
+                  title: fileName,
+                }
+              },
+            },
+          }),
+          defineArrayMember({
+            type: 'file',
+            name: 'filePDF',
+            title: 'PDF file block',
+            options: {
+              accept: 'application/pdf',
+            },
+            preview: {
+              select: {
+                tile: 'asset.originalFilename',
+              },
+            },
+          }),
+        ],
+      }),
+      defineField({
+        type: 'array',
+        name: 'bodyNormalized',
+        of: [
+          defineArrayMember({
+            type: 'block',
+            options: {
+              unstable_whitespaceOnPasteMode: 'normalize',
+            },
+          }),
+        ],
+      }),
+    ],
+  }),
+]
+
+function CopyPasteHarness({focusPath, document}: {focusPath?: Path; document?: SanityDocument}) {
+  return (
+    <TestWrapper schemaTypes={SCHEMA_TYPES}>
+      <TestForm document={document} focusPath={focusPath} />
+    </TestWrapper>
+  )
+}
 
 export type UpdateFn = () => {focusPath: Path; document: SanityDocument}
 
@@ -61,7 +140,7 @@ describe.skipIf(server.browser === 'webkit')('Portable Text Input', () => {
       const {getFocusedPortableTextEditor, insertPortableTextCopyPaste, waitForDocumentState} =
         testHelpers()
 
-      void render(<CopyPasteStory document={document} />)
+      void render(<CopyPasteHarness document={document} />)
 
       const $pte = await getFocusedPortableTextEditor('field-body')
 
@@ -85,7 +164,7 @@ describe.skipIf(server.browser === 'webkit')('Portable Text Input', () => {
       const {getFocusedPortableTextEditor, insertPortableTextCopyPaste, waitForDocumentState} =
         testHelpers()
 
-      void render(<CopyPasteStory document={document} />)
+      void render(<CopyPasteHarness document={document} />)
 
       const $pte = await getFocusedPortableTextEditor('field-bodyNormalized')
 
@@ -107,7 +186,7 @@ describe.skipIf(server.browser === 'webkit')('Portable Text Input', () => {
       const {getFocusedPortableTextEditor, insertPortableTextCopyPaste, waitForDocumentState} =
         testHelpers()
 
-      void render(<CopyPasteStory document={document} />)
+      void render(<CopyPasteHarness document={document} />)
 
       const $pte = await getFocusedPortableTextEditor('field-body')
 
@@ -135,7 +214,7 @@ describe.skipIf(server.browser === 'webkit')('Portable Text Input', () => {
     it.skipIf(server.browser === 'firefox')(`Added pasted image as a block`, async () => {
       const {getFocusedPortableTextEditor, pasteFileOverPortableTextEditor} = testHelpers()
 
-      void render(<CopyPasteStory document={document} />)
+      void render(<CopyPasteHarness document={document} />)
 
       const fileData = await loadTestFile('./static/dummy-image-1.jpg')
       const $pte = await getFocusedPortableTextEditor('field-body')
@@ -152,7 +231,7 @@ describe.skipIf(server.browser === 'webkit')('Portable Text Input', () => {
         hoverFileOverPortableTextEditor,
       } = testHelpers()
 
-      void render(<CopyPasteStory document={document} />)
+      void render(<CopyPasteHarness document={document} />)
 
       const fileData = await loadTestFile('./static/dummy-image-1.jpg')
       const $pte = await getFocusedPortableTextEditor('field-body')
@@ -174,7 +253,7 @@ describe.skipIf(server.browser === 'webkit')('Portable Text Input', () => {
     it(`Display error message on drag over if file is not accepted`, async () => {
       const {getFocusedPortableTextEditor, hoverFileOverPortableTextEditor} = testHelpers()
 
-      void render(<CopyPasteStory document={document} />)
+      void render(<CopyPasteHarness document={document} />)
 
       const fileData = await loadTestFile('./static/dummy.zip')
       const $pte = await getFocusedPortableTextEditor('field-body')
