--- packages/sanity/src/core/form/inputs/arrays/__tests__/ArrayInputStory.tsx
@@ -1,31 +1,0 @@
-import {defineField, defineType} from '@sanity/types'
-
-import {TestForm} from '../../../../../../test/browser/TestForm'
-import {TestWrapper} from '../../../../../../test/browser/TestWrapper'
-
-const SCHEMA_TYPES = [
-  defineType({
-    type: 'document',
-    name: 'test',
-    title: 'Test',
-    fields: [
-      defineField({
-        type: 'array',
-        name: 'tags',
-        title: 'Tags',
-        of: [{type: 'string'}],
-        options: {layout: 'tags'},
-      }),
-    ],
-  }),
-]
-
-function ArrayInputStory() {
-  return (
-    <TestWrapper schemaTypes={SCHEMA_TYPES}>
-      <TestForm />
-    </TestWrapper>
-  )
-}
-
-export default ArrayInputStory
